Refactored the grid properties into the grid namespace
