+++ packages/lytenyte-core/src/types/index.ts
@@ -2,7 +2,6 @@ export type * from "./api.js";
 export type * from "./events.js";
 export type * from "./column.js";
 export type * from "./grid.js";
-export type * from "./props.js";
 
 export type { ReactPlaceholderFn, DragItem, DragItemTransfer, DragItemSiteLocal } from "../dnd/types.js";
